requirements.txt
# Core
torch >= 2.1.0
torchvision >= 0.16
numpy == 1.26.2
scikit.learn == 1.3.2
pillow == 10.1.0
timm == 0.9.12
# Data Analysis/Logging
matplotlib == 3.8.2
pandas == 2.1.4
tensorboard == 2.15.1
tqdm == 4.66.1
wandb == 0.16.2
plotly == 5.18.0
rich == 13.7.0
# Config
click == 8.1.7
hydra-core == 1.3.2
python-dotenv == 1.0.0
# Notebook
ipykernel == 6.28.0
